acl manager( not completed yet)

+++ AclManager/src/Model/Table/AcosTable.php
@@ -44,7 +44,7 @@ class AcosTable extends SpiderTable
     public function validationDefault(Validator $validator)
     {
         $validator
-            ->add('id', 'valid', ['rule' => 'integer'])
+            ->add('id', 'valid', ['rule' => 'numeric'])
             ->allowEmpty('id', 'create');
         $validator
             ->allowEmpty('name');
@@ -55,7 +55,7 @@ class AcosTable extends SpiderTable
         $validator
             ->allowEmpty('model');
         $validator
-            ->add('foreign_key', 'valid', ['rule' => 'integer'])
+            ->add('foreign_key', 'valid', ['rule' => 'numeric'])
             ->allowEmpty('foreign_key');
         return $validator;
     }
