+++ AclManager/src/Model/Table/ArosAcosTable.php
@@ -46,7 +46,7 @@ class ArosAcosTable extends SpiderTable
     public function validationDefault(Validator $validator)
     {
         $validator
-            ->add('id', 'valid', ['rule' => 'integer'])
+            ->add('id', 'valid', ['rule' => 'numeric'])
             ->allowEmpty('id', 'create');
         return $validator;
     }
